winebus.sys: Map more SDL Xbox controllers
These device ids are usually returned by newer firmware versions.
Especially, the Android mode of the controller (when paired by bluetooth
to a Linux system) changes the input mapping of the controller. The
current kernel drivers do not correct this, xpadneo[1] is needed as the
driver to get correct mapping.

[1]: https://github.com/atar-axis/xpadneo

Signed-off-by: Kai Krakow <kai@kaishome.de>